+++ src/react/pages/GameDetailsPage.tsx
@@ -14,7 +14,7 @@ const GameDetailsPage: React.FC<GameDetailsPageProps> = ({game}: GameDetailsPage
 
   return <>
     <h1 style={{textAlign: 'center'}}>{game.name}</h1>
-    <div style={{ display: 'flex', flexDirection: 'row', justifyContent: 'space-between', flexWrap: 'wrap', gap: 35, padding: 25 }}>
+    <div style={{ display: 'flex', flexDirection: 'row', alignItems: 'start', justifyContent: 'space-between', flexWrap: 'wrap', gap: 35, padding: 25 }}>
       <GameCard game={game} onDownload={() => downloadManager.downloadGame(game.id)} verbose />
     </div>
   </>